+++ affinda/models/_models_py3.py
@@ -4764,7 +4764,7 @@ class Meta(msrest.serialization.Model):
     :vartype failed: bool
     :ivar expiry_time: The date/time in ISO-8601 format when the document will be automatically
      deleted.  Defaults to no expiry.
-    :vartype expiry_time: str
+    :vartype expiry_time: ~datetime.datetime
     """
 
     _validation = {
@@ -4779,7 +4779,7 @@ class Meta(msrest.serialization.Model):
         "ready": {"key": "ready", "type": "bool"},
         "ready_dt": {"key": "readyDt", "type": "iso-8601"},
         "failed": {"key": "failed", "type": "bool"},
-        "expiry_time": {"key": "expiryTime", "type": "str"},
+        "expiry_time": {"key": "expiryTime", "type": "iso-8601"},
     }
 
     def __init__(
@@ -4790,7 +4790,7 @@ class Meta(msrest.serialization.Model):
         failed: bool,
         file_name: Optional[str] = None,
         ready_dt: Optional[datetime.datetime] = None,
-        expiry_time: Optional[str] = None,
+        expiry_time: Optional[datetime.datetime] = None,
         **kwargs,
     ):
         """
@@ -4810,7 +4810,7 @@ class Meta(msrest.serialization.Model):
         :paramtype failed: bool
         :keyword expiry_time: The date/time in ISO-8601 format when the document will be automatically
          deleted.  Defaults to no expiry.
-        :paramtype expiry_time: str
+        :paramtype expiry_time: ~datetime.datetime
         """
         super(Meta, self).__init__(**kwargs)
         self.identifier = identifier
@@ -4844,7 +4844,7 @@ class InvoiceMeta(Meta, Components17Ashz6SchemasInvoicePropertiesMetaAllof1):
     :vartype failed: bool
     :ivar expiry_time: The date/time in ISO-8601 format when the document will be automatically
      deleted.  Defaults to no expiry.
-    :vartype expiry_time: str
+    :vartype expiry_time: ~datetime.datetime
     """
 
     _validation = {
@@ -4860,7 +4860,7 @@ class InvoiceMeta(Meta, Components17Ashz6SchemasInvoicePropertiesMetaAllof1):
         "ready": {"key": "ready", "type": "bool"},
         "ready_dt": {"key": "readyDt", "type": "iso-8601"},
         "failed": {"key": "failed", "type": "bool"},
-        "expiry_time": {"key": "expiryTime", "type": "str"},
+        "expiry_time": {"key": "expiryTime", "type": "iso-8601"},
     }
 
     def __init__(
@@ -4872,7 +4872,7 @@ class InvoiceMeta(Meta, Components17Ashz6SchemasInvoicePropertiesMetaAllof1):
         review_url: Optional[str] = None,
         file_name: Optional[str] = None,
         ready_dt: Optional[datetime.datetime] = None,
-        expiry_time: Optional[str] = None,
+        expiry_time: Optional[datetime.datetime] = None,
         **kwargs,
     ):
         """
@@ -4894,7 +4894,7 @@ class InvoiceMeta(Meta, Components17Ashz6SchemasInvoicePropertiesMetaAllof1):
         :paramtype failed: bool
         :keyword expiry_time: The date/time in ISO-8601 format when the document will be automatically
          deleted.  Defaults to no expiry.
-        :paramtype expiry_time: str
+        :paramtype expiry_time: ~datetime.datetime
         """
         super(InvoiceMeta, self).__init__(
             identifier=identifier,
@@ -5447,7 +5447,7 @@ class Paths1BwrvmkInvoicesPostRequestbodyContentMultipartFormDataSchema(
     :vartype language: str
     :ivar expiry_time: The date/time in ISO-8601 format when the document will be automatically
      deleted.  Defaults to no expiry.
-    :vartype expiry_time: str
+    :vartype expiry_time: ~datetime.datetime
     """
 
     _attribute_map = {
@@ -5457,7 +5457,7 @@ class Paths1BwrvmkInvoicesPostRequestbodyContentMultipartFormDataSchema(
         "url": {"key": "url", "type": "str"},
         "wait": {"key": "wait", "type": "bool"},
         "language": {"key": "language", "type": "str"},
-        "expiry_time": {"key": "expiryTime", "type": "str"},
+        "expiry_time": {"key": "expiryTime", "type": "iso-8601"},
     }
 
     def __init__(
@@ -5469,7 +5469,7 @@ class Paths1BwrvmkInvoicesPostRequestbodyContentMultipartFormDataSchema(
         url: Optional[str] = None,
         wait: Optional[bool] = True,
         language: Optional[str] = None,
-        expiry_time: Optional[str] = None,
+        expiry_time: Optional[datetime.datetime] = None,
         **kwargs,
     ):
         """
@@ -5491,7 +5491,7 @@ class Paths1BwrvmkInvoicesPostRequestbodyContentMultipartFormDataSchema(
         :paramtype language: str
         :keyword expiry_time: The date/time in ISO-8601 format when the document will be automatically
          deleted.  Defaults to no expiry.
-        :paramtype expiry_time: str
+        :paramtype expiry_time: ~datetime.datetime
         """
         super(Paths1BwrvmkInvoicesPostRequestbodyContentMultipartFormDataSchema, self).__init__(
             **kwargs
@@ -5653,7 +5653,7 @@ class Paths7EskthResumesPostRequestbodyContentMultipartFormDataSchema(msrest.ser
     :vartype language: str
     :ivar expiry_time: The date/time in ISO-8601 format when the document will be automatically
      deleted.  Defaults to no expiry.
-    :vartype expiry_time: str
+    :vartype expiry_time: ~datetime.datetime
     """
 
     _attribute_map = {
@@ -5663,7 +5663,7 @@ class Paths7EskthResumesPostRequestbodyContentMultipartFormDataSchema(msrest.ser
         "url": {"key": "url", "type": "str"},
         "wait": {"key": "wait", "type": "bool"},
         "language": {"key": "language", "type": "str"},
-        "expiry_time": {"key": "expiryTime", "type": "str"},
+        "expiry_time": {"key": "expiryTime", "type": "iso-8601"},
     }
 
     def __init__(
@@ -5675,7 +5675,7 @@ class Paths7EskthResumesPostRequestbodyContentMultipartFormDataSchema(msrest.ser
         url: Optional[str] = None,
         wait: Optional[bool] = True,
         language: Optional[str] = None,
-        expiry_time: Optional[str] = None,
+        expiry_time: Optional[datetime.datetime] = None,
         **kwargs,
     ):
         """
@@ -5697,7 +5697,7 @@ class Paths7EskthResumesPostRequestbodyContentMultipartFormDataSchema(msrest.ser
         :paramtype language: str
         :keyword expiry_time: The date/time in ISO-8601 format when the document will be automatically
          deleted.  Defaults to no expiry.
-        :paramtype expiry_time: str
+        :paramtype expiry_time: ~datetime.datetime
         """
         super(Paths7EskthResumesPostRequestbodyContentMultipartFormDataSchema, self).__init__(
             **kwargs
@@ -5750,7 +5750,7 @@ class Paths8DdhfcRedactedResumesPostRequestbodyContentMultipartFormDataSchema(
     :vartype redact_gender: bool
     :ivar expiry_time: The date/time in ISO-8601 format when the document will be automatically
      deleted.  Defaults to no expiry.
-    :vartype expiry_time: str
+    :vartype expiry_time: ~datetime.datetime
     """
 
     _attribute_map = {
@@ -5768,7 +5768,7 @@ class Paths8DdhfcRedactedResumesPostRequestbodyContentMultipartFormDataSchema(
         "redact_locations": {"key": "redactLocations", "type": "bool"},
         "redact_dates": {"key": "redactDates", "type": "bool"},
         "redact_gender": {"key": "redactGender", "type": "bool"},
-        "expiry_time": {"key": "expiryTime", "type": "str"},
+        "expiry_time": {"key": "expiryTime", "type": "iso-8601"},
     }
 
     def __init__(
@@ -5788,7 +5788,7 @@ class Paths8DdhfcRedactedResumesPostRequestbodyContentMultipartFormDataSchema(
         redact_locations: Optional[bool] = True,
         redact_dates: Optional[bool] = True,
         redact_gender: Optional[bool] = True,
-        expiry_time: Optional[str] = None,
+        expiry_time: Optional[datetime.datetime] = None,
         **kwargs,
     ):
         """
@@ -5826,7 +5826,7 @@ class Paths8DdhfcRedactedResumesPostRequestbodyContentMultipartFormDataSchema(
         :paramtype redact_gender: bool
         :keyword expiry_time: The date/time in ISO-8601 format when the document will be automatically
          deleted.  Defaults to no expiry.
-        :paramtype expiry_time: str
+        :paramtype expiry_time: ~datetime.datetime
         """
         super(
             Paths8DdhfcRedactedResumesPostRequestbodyContentMultipartFormDataSchema, self
@@ -6025,7 +6025,7 @@ class PathsYlw96JobDescriptionsPostRequestbodyContentMultipartFormDataSchema(
     :vartype language: str
     :ivar expiry_time: The date/time in ISO-8601 format when the document will be automatically
      deleted.  Defaults to no expiry.
-    :vartype expiry_time: str
+    :vartype expiry_time: ~datetime.datetime
     """
 
     _attribute_map = {
@@ -6035,7 +6035,7 @@ class PathsYlw96JobDescriptionsPostRequestbodyContentMultipartFormDataSchema(
         "url": {"key": "url", "type": "str"},
         "wait": {"key": "wait", "type": "bool"},
         "language": {"key": "language", "type": "str"},
-        "expiry_time": {"key": "expiryTime", "type": "str"},
+        "expiry_time": {"key": "expiryTime", "type": "iso-8601"},
     }
 
     def __init__(
@@ -6047,7 +6047,7 @@ class PathsYlw96JobDescriptionsPostRequestbodyContentMultipartFormDataSchema(
         url: Optional[str] = None,
         wait: Optional[bool] = True,
         language: Optional[str] = None,
-        expiry_time: Optional[str] = None,
+        expiry_time: Optional[datetime.datetime] = None,
         **kwargs,
     ):
         """
@@ -6069,7 +6069,7 @@ class PathsYlw96JobDescriptionsPostRequestbodyContentMultipartFormDataSchema(
         :paramtype language: str
         :keyword expiry_time: The date/time in ISO-8601 format when the document will be automatically
          deleted.  Defaults to no expiry.
-        :paramtype expiry_time: str
+        :paramtype expiry_time: ~datetime.datetime
         """
         super(
             PathsYlw96JobDescriptionsPostRequestbodyContentMultipartFormDataSchema, self